# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TradeDesk-AI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/orders/preview": {
      "post": {
        "summary": "Preview Order",
        "description": "Build an order card from a structured intent (the same object the LLM emits).\n\nNothing is sent. The card has to be approved separately, and the hash on it is what the\napprove call must echo. Ambiguous or unknown instruments come back as a question.",
        "operationId": "preview_order_api_orders_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderIntent-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatReply"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/approvals/{pending_id}/approve": {
      "post": {
        "summary": "Approve",
        "operationId": "approve_api_approvals__pending_id__approve_post",
        "parameters": [
          {
            "name": "pending_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pending Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResult"
                }
              }
            }
          },
          "404": {
            "description": "Unknown approval id"
          },
          "409": {
            "description": "Nothing was sent: stale, expired, voided, blocked or re-quoted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalConflict"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/approvals/{pending_id}/reject": {
      "post": {
        "summary": "Reject",
        "operationId": "reject_api_approvals__pending_id__reject_post",
        "parameters": [
          {
            "name": "pending_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pending Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingOrder"
                }
              }
            }
          },
          "404": {
            "description": "Unknown approval id"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalConflict"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/preview": {
      "post": {
        "summary": "Preview Plan",
        "description": "Build a plan card from structured steps (the same object the assistant uses). Nothing is sent.",
        "operationId": "preview_plan_api_plans_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatReply"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/approve": {
      "post": {
        "summary": "Approve Plan",
        "description": "Approve a whole plan. The steps then run in order in the background, each through the same\nexecutor as a single order; follow progress on the WebSocket or with GET /plans/{id}/report.",
        "operationId": "approve_plan_api_plans__plan_id__approve_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanApproveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanReport"
                }
              }
            }
          },
          "404": {
            "description": "Unknown plan id"
          },
          "409": {
            "description": "Nothing was run: stale, expired, voided, blocked or re-quoted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalConflict"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/reject": {
      "post": {
        "summary": "Reject Plan",
        "operationId": "reject_plan_api_plans__plan_id__reject_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "404": {
            "description": "Unknown plan id"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalConflict"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/report": {
      "get": {
        "summary": "Plan Report",
        "description": "Where each step stands: filled, partly filled, rejected, not sent. Nothing is hidden.",
        "operationId": "plan_report_api_plans__plan_id__report_get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanReport"
                }
              }
            }
          },
          "404": {
            "description": "Unknown plan id"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules": {
      "get": {
        "summary": "Get Rules",
        "description": "Newest first. A rule fires once; fired and cancelled rules stay listed.",
        "operationId": "get_rules_api_rules_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RuleStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rule"
                  },
                  "title": "Response Get Rules Api Rules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Rule",
        "description": "Create a standing instruction from structured fields (the same object the assistant uses).\n\nA rule never sends an order: when it fires it prepares an approval card or an alert.",
        "operationId": "create_rule_api_rules_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRuleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatReply"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/{rule_id}": {
      "delete": {
        "summary": "Delete Rule",
        "operationId": "delete_rule_api_rules__rule_id__delete",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            }
          },
          "404": {
            "description": "Unknown rule"
          },
          "409": {
            "description": "Rule is not active"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit": {
      "get": {
        "summary": "Get Audit",
        "description": "Newest first.",
        "operationId": "get_audit_api_audit_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AuditKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEvent"
                  },
                  "title": "Response Get Audit Api Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chat": {
      "post": {
        "summary": "Chat",
        "description": "Ask the copilot. Order requests come back as cards; nothing is ever sent from here.",
        "operationId": "chat_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatReply"
                }
              }
            }
          },
          "503": {
            "description": "Assistant or broker unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApproveRequest": {
        "properties": {
          "order_hash": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Order Hash"
          },
          "acknowledgment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledgment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "order_hash"
        ],
        "title": "ApproveRequest",
        "description": "The client must echo the hash of the exact card it is approving."
      },
...
      "ChatReply": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "cards": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/PendingOrderCard"
                },
                {
                  "$ref": "#/components/schemas/PlanCard"
                },
                {
                  "$ref": "#/components/schemas/RuleCard"
                },
                {
                  "$ref": "#/components/schemas/AmbiguityCard"
                },
                {
                  "$ref": "#/components/schemas/NoticeCard"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "ambiguity": "#/components/schemas/AmbiguityCard",
                  "notice": "#/components/schemas/NoticeCard",
                  "pending_order": "#/components/schemas/PendingOrderCard",
                  "plan": "#/components/schemas/PlanCard",
                  "rule": "#/components/schemas/RuleCard"
                }
              }
            },
            "type": "array",
            "title": "Cards"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "cards"
        ],
        "title": "ChatReply"
      },
...
      "ExecutionResult": {
        "properties": {
          "pending": {
            "$ref": "#/components/schemas/PendingOrder"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "SENT",
              "REJECTED",
              "UNKNOWN"
            ],
            "title": "Outcome"
          },
          "order": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Order"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pending",
          "outcome",
          "order",
          "message"
        ],
        "title": "ExecutionResult",
        "description": "What happened after an approved card was sent.\n\nSENT      the broker accepted it (see `order` for its live status)\nREJECTED  the broker refused it (`order.rejection_reason` says why)\nUNKNOWN   the request timed out and we could not confirm either way. The order is\n          never re-sent; it is reconciled against the broker's order book."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderIntent-Input": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/OrderAction"
          },
          "instrument_ref": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument Ref"
          },
          "option": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptionRef-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "future": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FutureRef-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "side": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Side"
              },
              {
                "type": "null"
              }
            ]
          },
          "quantity": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "lots": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Lots"
          },
          "amount_paise": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Paise"
          },
          "fraction_of_holding": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Fraction Of Holding"
          },
          "order_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrderType"
              },
              {
                "type": "null"
              }
            ]
          },
          "limit_price": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Price"
          },
          "trigger_price": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Price"
          },
          "product": {
            "$ref": "#/components/schemas/Product",
            "default": "CNC"
          },
          "validity": {
            "$ref": "#/components/schemas/Validity",
            "default": "DAY"
          },
          "target_order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Order Id"
          },
          "via_voice": {
            "type": "boolean",
            "title": "Via Voice",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action"
        ],
        "title": "OrderIntent",
        "description": "The only order-shaped object the LLM may emit. Not an order: code resolves and validates it."
      },
...
      "RuleStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "FIRED",
          "CANCELLED"
        ],
        "title": "RuleStatus"
      },
...
      "Side": {
        "type": "string",
        "enum": [
          "BUY",
          "SELL"
        ],
        "title": "Side"
      },
...
    }
  }
}